Added docker workflow help packages in emacs. Inspired b http://manuel-uberti.github.io/emacs/2017/10/19/docker/
Packages addded -
1. dockerfile-mode
2. docker-compose-mode
3. docker-tramp
4. eshell-bookmark

The first two packages add syntax highlighting, completion and a bunch
of useful key bindings to work with Dockerfile and docker-compose.yml.

The real magic, though, happens with docker-tramp. The beauty of
interacting directly with the contents of a container is
impressive. Combine docker-tramp with the handy completion of counsel
and the power of Dired and you might feel like Henry Dorsett Case in
his ecstatic hunt for the Neuromancer.